refactor and move code to experimental.units
